move attach modals to app/components
