+++ app/pages/project/instances/instance/tabs/NetworkingTab.tsx
@@ -21,12 +21,12 @@ import {
 } from '@oxide/api'
 import { IpGlobal24Icon, Networking24Icon } from '@oxide/design-system/icons/react'
 
+import { AttachEphemeralIpModal } from '~/components/AttachEphemeralIpModal'
+import { AttachFloatingIpModal } from '~/components/AttachFloatingIpModal'
 import { HL } from '~/components/HL'
 import { CreateNetworkInterfaceForm } from '~/forms/network-interface-create'
 import { EditNetworkInterfaceForm } from '~/forms/network-interface-edit'
 import { getInstanceSelector, useInstanceSelector, useProjectSelector } from '~/hooks'
-import { AttachEphemeralIpModal } from '~/pages/project/floating-ips/AttachEphemeralIpModal'
-import { AttachFloatingIpModal } from '~/pages/project/floating-ips/AttachFloatingIpModal'
 import { confirmAction } from '~/stores/confirm-action'
 import { confirmDelete } from '~/stores/confirm-delete'
 import { addToast } from '~/stores/toast'
